Improve error checks in processTransfers
